# Fixtures/Sources/WireOpenAPIBootstrapExample/openapi.yaml
openapi: 3.1.0
info:
  title: Tasks
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /tasks/{id}:
    get:
      operationId: getTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
  /tasks:
    get:
      operationId: listTasks
      responses:
        '500': { description: Failed }
        '200':
          description: Task ids
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /tasks/{id}/summary:
    get:
      operationId: summariseTask
      parameters:
        # Declared by reference, not inline. Ordinary OpenAPI, and the shape that exposed reading the
        # document as a dictionary: a `$ref` carries no `name`, so it was dropped and a handler binding
        # `id` was rejected for binding something "the document does not declare". OpenAPIKit resolves it.
        - { $ref: '#/components/parameters/TaskId' }
        - { name: include-done, in: query, required: false, schema: { type: boolean } }
        - { name: X-Request-Id, in: header, required: false, schema: { type: string } }
      responses:
        '200':
          description: A task summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: No such task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403': { description: Not allowed }
  /tasks/new:
    post:
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Task' }
      parameters:
        - { name: title, in: query, required: true, schema: { type: string } }
      responses:
        '500': { description: Failed }
        # Declared so the DecodingError mapping has a documented case to construct *inside* the
        # forwarder. Without it the mapping still answers, from the terminal — the difference is only
        # which side of the generator the response is built on.
        '422': { description: Could not decode }
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
  /tasks/{id}/delete:
    post:
      operationId: deleteTask
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '500': { description: Failed }
        '202': { description: Accepted }
        '204': { description: Deleted }
  /tasks/{id}/replace:
    post:
      operationId: replaceTask
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Task' }
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: title, in: query, required: true, schema: { type: string } }
      responses:
        '500': { description: Failed }
        '200':
          description: Replaced
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
components:
  parameters:
    TaskId: { name: id, in: path, required: true, schema: { type: string } }
  schemas:
    Problem:
      type: object
      required: [message]
      properties:
        message: { type: string }
    Task:
      type: object
      required: [id, title]
      properties:
        id:
          type: string
        title:
          type: string